Split Chinese language into Simplified script and Traditional script
There are two kinds of Chinese characters, Simplified Chinese (zh-Hans or zh-CN) and Traditional Chinese (zh-Hant or zh-HK, zh-TW, zh-MO etc.). The original "zh.ini" of _fallback theme uses Traditional Chinese, as I'm not sure the region of it (their typeface look the same), I renamed it "zh-Hant.ini" in order to represent Traditional Chinese.
I did not modify the content of the file. Only renamed it.